Move classes into the lib directory
