+++ lib/irc.js
